Allow output_type to work as advertised.
According to the readme, specifying an output_type will allow stats
to pass to a configured endpoint by the same name. However, this
next statement prevented anything except graphite or opentsdb
output_types from ever being queued and eventually posted.

This change fixes this behavior so it works as documented.